# swagger-v2.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Products API",
    "description": "Manage products, categories, and reviews",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/api/v2"
    }
  ],
  "paths": {
    "/categories": {
      "get": {
        "tags": ["categories"],
        "responses": {
          "200": {
          "description": "OK",
          "content": {
              "application/json": {
              "schema": {
                  "type": "array",
                  "items": {
                  "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["categories"],
        "requestBody": {
            "content": {
            "application/json": {
                "schema": {
                "$ref": "#/components/schemas/Category"
                }
            }
            },
            "required": true
        },
        "responses": {
            "200": {
            "description": "OK",
            "content": {
                "application/json": {
                "schema": {
                    "$ref": "#/components/schemas/Category"
                }
                }
            }
            },
            "400": {
            "description": "Invalid input"
            }
        }
      }
    },
    "/categories/{category_id}": {
      "parameters": [
        {
          "in": "path",
          "name": "category_id",
          "required": true,
          "schema": {
          "type": "integer"
          }
        }
      ],
      "get": {
        "tags": ["categories"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "tags": ["categories"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": ["categories"],
        "responses": {
          "204": {
            "description": "Deleted successfully"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/products": {
      "parameters": [
        {
          "in": "query",
          "name": "page",
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "tags": ["products"],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": [
                "date",
                "price",
                "rating"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["products"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/products/{product_id}": {
      "parameters": [
        {
          "in": "path",
          "name": "product_id",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "tags": ["products"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "tags": ["products"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "tags": ["products"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "quantity": {
                    "type": "integer",
                    "example": 2
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": ["products"],
        "responses": {
          "204": {
            "description": "Deleted successfully"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/products/{product_id}/reviews": {
      "parameters": [
        {
          "in": "path",
          "name": "product_id",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "tags": ["products", "reviews"],
        "parameters": [
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": [
                "date",
                "votes",
                "rating"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "tags": ["products", "reviews"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Review"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/reviews/{review_id}": {
      "parameters": [
        {
          "in": "path",
          "name": "review_id",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "tags": ["reviews"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "tags": ["reviews"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Review"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "tags": ["reviews"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "helpful_votes": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": ["reviews"],
        "responses": {
          "204": {
            "description": "Deleted successfully"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Category": {
        "properties": {
          "id": {
            "type": "integer",
            "example": 1,
            "readOnly": true
          },
          "name": {
            "type": "string",
            "example": "Electronics"
          },
          "icon": {
            "type": "string",
            "example": "💻"
          }
        }
      },
      "Product": {
        "properties": {
          "id": {
            "type": "integer",
            "example": 1,
            "readOnly": true
          },
          "name": {
            "type": "string",
            "example": "Laptop"
          },
          "description": {
            "type": "string",
            "example": "Surf the web and code all on the same machine!"
          },
          "price": {
            "type": "number",
            "example": 399.99
          },
          "quantity": {
            "type": "integer",
            "example": 50
          },
          "rating": {
            "type": "number",
            "example": 4.5,
            "readOnly": true
          },
          "product_date": {
            "type": "string",
            "example": "2020-01-01",
            "readOnly": true
          },
          "category_id": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "Review": {
        "properties": {
          "id": {
            "type": "integer",
            "example": 1,
            "readOnly": true
          },
          "title": {
            "type": "string",
            "example": "This product is the best!"
          },
          "review_text": {
            "type": "string",
            "example": "I really like how it does the thing it's supposed to do."
          },
          "rating": {
            "type": "integer",
            "example": 5
          },
          "review_date": {
            "type": "string",
            "example": "2020-01-01",
            "readOnly": true
          },
          "helpful_votes": {
            "type": "integer",
            "example": 30,
            "readOnly": true
          },
          "product_id": {
            "type": "integer",
            "example": 1,
            "readOnly": true
          }
        }
      }
    }
  }
}